restructured the game loop
