+++ internal/server/Governance.go
@@ -18,22 +18,27 @@ func (s *Server) RunRulerAction(bike objects.IMegaBike) uuid.UUID {
 }
 
 func (s *Server) RulerElection(agents []objects.IBaseBiker, governance utils.Governance) uuid.UUID {
-	votes := make([]voting.IdVoteMap, len(agents))
-	for i, agent := range agents {
+	votes := make(map[uuid.UUID]voting.IdVoteMap)
+	for _, agent := range agents {
 		switch governance {
 		case utils.Dictatorship:
-			votes[i] = agent.VoteDictator()
+			votes[agent.GetID()] = agent.VoteDictator()
 		case utils.Leadership:
-			votes[i] = agent.VoteLeader()
+			votes[agent.GetID()] = agent.VoteLeader()
 		}
 	}
 
-	IVotes := make([]voting.IVoter, len(votes))
+	IVotes := make(map[uuid.UUID]voting.IVoter, len(votes))
 	for i, vote := range votes {
 		IVotes[i] = vote
 	}
 
-	ruler := voting.WinnerFromDist(IVotes)
+	// all agents have equal voting power in the election
+	weights := make(map[uuid.UUID]float64)
+	for _, agent := range agents {
+		weights[agent.GetID()] = 1.0
+	}
+	ruler := voting.WinnerFromDist(IVotes, weights)
 	return ruler
 }
 
@@ -41,7 +46,6 @@ func (s *Server) RunDemocraticAction(bike objects.IMegaBike, weights map[uuid.UU
 	// map of the proposed lootboxes by bike (for each bike a list of lootbox proposals is made, with one lootbox proposed by each agent on the bike)
 	agents := bike.GetAgents()
 	proposedDirections := make([]uuid.UUID, len(agents))
-
 	// the direction proposal process is the same for both types of democracies
 	for i, agent := range agents {
 		// agents that have decided to stay on the bike (and that haven't been kicked off it)
@@ -53,12 +57,12 @@ func (s *Server) RunDemocraticAction(bike objects.IMegaBike, weights map[uuid.UU
 	}
 
 	// pass the pitched directions of a bike to all agents on that bike and get their final vote
-	finalVotes := make([]voting.LootboxVoteMap, len(agents))
+	finalVotes := make(map[uuid.UUID]voting.LootboxVoteMap)
 
 	// pass the pitched directions of a bike to all agents on that bike and get their final vote
-	for i, agent := range agents {
+	for _, agent := range agents {
 		// ---------------------------VOTING ROUTINE - STEP 2 ---------------------
-		finalVotes[i] = agent.FinalDirectionVote(proposedDirections)
+		finalVotes[agent.GetID()] = agent.FinalDirectionVote(proposedDirections)
 	}
 
 	// ---------------------------VOTING ROUTINE - STEP 3 --------------
